content: move some logic to buttons.go
